requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
sqlalchemy==2.0.28
psycopg2-binary==2.9.9
pyyaml==6.0.1
python-dotenv==1.0.1
schedule==1.2.1
tenacity==8.2.3 